KELAS12/CODING/A/08-1018/pyproject.toml
[project]
name = "08-1018"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "python-dotenv>=1.1.1",
]
